mesh-watcher: add mesh-watcher
The network may crash on the SoCs, but the SoC itself does not. This
leads to a node no longer being accessible in a mesh network.  If the
node is placed in a location that is not easily accessible, e.g. in a
high tower, it can cause a lot of problems. Therefore we check the
link-local connectivity on the configured mesh interfaces. For now, only
babeld is supported.

Signed-off-by: Nick Hainke <vincent@systemli.org>